requirements.txt
streamlit
google-cloud-firestore
google-cloud